Gradle run works
