+++ examples/visual-effects/README.md
@@ -1,1 +1,1 @@
-Several visual effects implmented with Compose Multiplatform, used in 1.0 release video.
+Several visual effects implmented with Compose Multiplatform, used in 1.0 release announce video.
