+++ examples/visual-effects/build.gradle.kts
@@ -6,7 +6,7 @@ plugins {
     // __KOTLIN_COMPOSE_VERSION__
     kotlin("jvm") version "1.5.31"
     // __LATEST_COMPOSE_RELEASE_VERSION__
-    id("org.jetbrains.compose") version "1.0.0-rc12"
+    id("org.jetbrains.compose") version "1.0.0"
 }
 
 group = "me.user"
@@ -22,18 +22,36 @@ dependencies {
     implementation(compose.desktop.currentOs)
 }
 
-tasks.withType<KotlinCompile>() {
+tasks.withType<KotlinCompile> {
     kotlinOptions.jvmTarget = "11"
     kotlinOptions.freeCompilerArgs += "-Xopt-in=kotlin.RequiresOptIn"
 }
 
 compose.desktop {
     application {
-        mainClass = "org.jetbrains.compose.demo.words.MainKt"
+        mainClass = "org.jetbrains.compose.demo.visuals.MainKt"
         nativeDistributions {
             targetFormats(TargetFormat.Dmg, TargetFormat.Msi, TargetFormat.Deb)
             packageName = "compose-demo"
             packageVersion = "1.0.0"
         }
+    }
+}
+
+afterEvaluate {
+    val additionalArguments = mutableListOf<String>()
+
+    val runTask = tasks.named<JavaExec>("run") {
+        this.args = additionalArguments
+    }
+
+    tasks.register("runWords") {
+        additionalArguments.add("words")
+        dependsOn(runTask)
+    }
+
+    tasks.register("runWave") {
+        additionalArguments.add("wave")
+        dependsOn(runTask)
     }
 }
